pyproject.toml

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
[tool.poetry]
name = "ai-affiliate-engine"
version = "0.1.0"
description = ""
authors = ["JP"]
package-mode = false

[tool.poetry.dependencies]
pillow = ">=10.0.0,<11.0.0"
python = ">=3.11,<3.13"
openai = ">=1.12.0,<2.0.0"
pydantic = ">=2.5.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
python-dateutil = ">=2.8.2,<3.0.0"
pyyaml = "^6.0.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate-post = "cli:generate_post"
content-factory = "content_factory.cli:main"


